requirements.txt
torch==2.5.1
torchvision==0.20.1
fastapi>=0.115.0
uvicorn[standard]>=0.33.0
python-multipart>=0.0.19
gradio>=6.12.0
numpy>=1.26.0
Pillow>=10.3.0
opencv-python>=4.10.0.84
pytesseract>=0.3.10
pytest>=7.4.0